refactor: user package structure
- divide package by responsibility
