+++ src/main/java/com/dailywriting/web/user/controller/TokenController.java
@@ -1,8 +1,11 @@
-package com.dailywriting.web.user;
+package com.dailywriting.web.user.controller;
 
 import com.dailywriting.web.common.CommonExceptionResponseBody;
 import com.dailywriting.web.security.JwtPayload;
 import com.dailywriting.web.security.JwtTokenProvider;
+import com.dailywriting.web.user.exception.LoginFailException;
+import com.dailywriting.web.user.domain.User;
+import com.dailywriting.web.user.domain.UserService;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
