+++ src/main/java/com/dailywriting/web/user/domain/UserRepository.java
@@ -1,4 +1,4 @@
-package com.dailywriting.web.user;
+package com.dailywriting.web.user.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
