+++ src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -4,6 +4,11 @@
       "name": "app.jwt.secret-key",
       "type": "java.lang.String",
       "description": "Description for app.jwt.secret-key."
+    },
+    {
+      "name": "app.jwt.expiration-seconds",
+      "type": "java.lang.Long",
+      "description": "Description for app.jwt.expiration-seconds."
     }
   ]
 }
