+++ src/test/java/com/dailywriting/web/user/UserServiceUnitTest.java
@@ -1,5 +1,9 @@
 package com.dailywriting.web.user;
 
+import com.dailywriting.web.user.domain.User;
+import com.dailywriting.web.user.domain.UserRepository;
+import com.dailywriting.web.user.domain.UserService;
+import com.dailywriting.web.user.exception.UserDuplicateException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.security.crypto.password.PasswordEncoder;
